userGreeting in resultsPage | moved domain, subdomain, previewQuestions to /components/results
