+++ medpass/src/app/dashboard/results/page.tsx
@@ -3,35 +3,20 @@
 import { useState, useEffect } from 'react';
 import Domain from '@/components/results/domain';
 import Sidebar from '@/components/navbar';
-import { Performance, ExamResult } from '@/types/domain';
-import { useSession } from 'next-auth/react'; // Import session hook
-
-// Define the API response type
-interface ApiResponseItem {
-    ExamResults: {
-        ExamResultsID: number;
-        StudentID?: number;
-        StudentName?: string;
-        ExamID?: number;
-        ExamName?: string;
-        Score?: number;
-        PassOrFail?: boolean;
-        Timestamp?: string | null;
-        ClerkshipID?: number | null;
-    };
-    Performances: Performance[];
-}
+import UserGreeting from '@/components/UserGreeting';
+import { ApiResponseItem, ExamResult } from '@/types/results';
+import { useSession } from 'next-auth/react';
 
 export default function DomainPage() {
-    const { data: session } = useSession(); // Get authentication session
+    const { data: session } = useSession();
     const [performanceData, setPerformanceData] = useState<ApiResponseItem[] | null>(null);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState<string | null>(null);
 
     useEffect(() => {
-        if (!session?.accessToken) return; // Return early if no access token
+        if (!session?.accessToken) return;
         fetchData();
-    }, [session]); // Re-run when session changes
+    }, [session]);
 
     const fetchData = async () => {
         setLoading(true);
@@ -64,14 +49,20 @@ export default function DomainPage() {
     };
 
     return (
-        <div className="min-h-screen bg-gray-900 flex">
+        <div className="min-h-screen bg-gray-900 text-slate-100 p-4">
             <Sidebar />
 
-            <div className="flex-1 p-6 max-w-4xl mx-auto text-white">
-                <div className="mb-6">
-                    <h1 className="text-2xl font-bold mb-4">Student Historical Performance</h1>
-                </div>
+            {/* Progress header */}
+            <div className="flex justify-between items-center mb-6 px-2">
+                <h1 className="text-2xl font-bold ml-20">Student Historical Performance</h1>
 
+                {/* Right side: User info */}
+                <div className="flex items-center">
+                    <UserGreeting className="text-xl font-bold" />
+                </div>
+            </div>
+
+            <div className="flex-1 max-w-4xl mx-auto">
                 {!session?.accessToken && (
                     <div className="bg-yellow-500/20 border border-yellow-500 text-yellow-300 px-4 py-3 rounded mb-4">
                         <p>Please sign in to view your performance data.</p>
