+++ medpass/src/lib/resultsUtils.ts
